Вторая правка валидации кол-ва чисел в цене после запятой

+++ backend/src/Lighthouse/CoreBundle/DataTransformer/MoneyModelTransformer.php
@@ -60,7 +60,7 @@ class MoneyModelTransformer implements DataTransformerInterface
     {
         if (null !== $value && '' !== $value) {
             $value = $value * $this->divider;
-            $value = (int) (string) $value;
+            $value = (float) (string) $value;
         }
         return new Money($value);
     }
